删除category时检查是否已有文章关联此tag

+++ src/main/java/io/github/jiashunx/app/miniblog/console/CategoryManageServlet.java
@@ -1,7 +1,10 @@
 package io.github.jiashunx.app.miniblog.console;
 
 import com.jfinal.kit.Kv;
+import io.github.jiashunx.app.miniblog.exception.MiniBlogException;
+import io.github.jiashunx.app.miniblog.model.entity.ArticleCategoryEntity;
 import io.github.jiashunx.app.miniblog.model.entity.CategoryEntity;
+import io.github.jiashunx.app.miniblog.service.ArticleCategoryService;
 import io.github.jiashunx.app.miniblog.service.CategoryService;
 import io.github.jiashunx.app.miniblog.util.BlogUtils;
 import io.github.jiashunx.masker.rest.framework.MRestRequest;
@@ -25,9 +28,11 @@ public class CategoryManageServlet extends AbstractRestServlet {
     private static final String CATEGORY_MANAGE_HTML = IOUtils.loadContentFromClasspath("template/console/category-index.html");
 
     private final CategoryService categoryService;
+    private final ArticleCategoryService articleCategoryService;
 
-    public CategoryManageServlet(CategoryService categoryService) {
+    public CategoryManageServlet(CategoryService categoryService, ArticleCategoryService articleCategoryService) {
         this.categoryService = Objects.requireNonNull(categoryService);
+        this.articleCategoryService = Objects.requireNonNull(articleCategoryService);
     }
 
     @PostMapping(url = "/console/category/create")
@@ -51,6 +56,10 @@ public class CategoryManageServlet extends AbstractRestServlet {
     @PostMapping(url = "/console/category/delete")
     public void delete(MRestRequest request, MRestResponse response) {
         CategoryEntity entity = request.parseBodyToObj(CategoryEntity.class);
+        List<ArticleCategoryEntity> articleCategoryEntityList = articleCategoryService.listCategoryArticles(entity.getCategoryId());
+        if (articleCategoryEntityList.isEmpty()) {
+            throw new MiniBlogException(String.format("there is %d articles linked to category: %s", articleCategoryEntityList.size(), entity.getCategoryName()));
+        }
         categoryService.deleteById(entity.getCategoryId());
     }
